support collapsed editor for single value segment
